refactor(workflows): extract engine runtime chokepoint
Assistant-model: OpenAI GPT-5
