+++ packages/workflows/src/authoring/workflow.ts
@@ -54,17 +54,27 @@ export type WorkflowProvidedInputsFromSchemas<TSchemas extends WorkflowInputSche
     }[SchemaKeys<TSchemas>]>>
 ) & WorkflowInputValues;
 
-export type WorkflowOutputsFromSchemas<TSchemas extends WorkflowOutputSchemaMap> = (
-  [SchemaKeys<TSchemas>] extends [never]
-    ? {}
-    : Simplify<UnionToIntersection<{
-      readonly [K in SchemaKeys<TSchemas>]: DeclaredOutputEntry<K, TSchemas[K]>;
-    }[SchemaKeys<TSchemas>]>>
-) & WorkflowOutputValues;
+type WorkflowDeclaredOutputsFromSchemas<TSchemas extends WorkflowOutputSchemaMap> = [SchemaKeys<TSchemas>] extends [never]
+  ? {}
+  : Simplify<UnionToIntersection<{
+    readonly [K in SchemaKeys<TSchemas>]: DeclaredOutputEntry<K, TSchemas[K]>;
+  }[SchemaKeys<TSchemas>]>>;
+
+export type WorkflowOutputsFromSchemas<TSchemas extends WorkflowOutputSchemaMap> =
+  WorkflowDeclaredOutputsFromSchemas<TSchemas> & WorkflowOutputValues;
+
+type NoExtraWorkflowOutputs<TDeclared, TActual extends TDeclared> = TActual &
+  Record<Exclude<keyof TActual, keyof TDeclared>, never>;
+
+type WorkflowRunOutputResult<
+  TOutputs extends WorkflowOutputSchemaMap,
+  TActualOutputs extends WorkflowDeclaredOutputsFromSchemas<TOutputs>,
+> = NoExtraWorkflowOutputs<WorkflowDeclaredOutputsFromSchemas<TOutputs>, TActualOutputs>;
 
 export interface AuthoredWorkflowSpec<
   TInputs extends WorkflowInputSchemaMap,
   TOutputs extends WorkflowOutputSchemaMap,
+  TActualOutputs extends WorkflowDeclaredOutputsFromSchemas<TOutputs> = WorkflowDeclaredOutputsFromSchemas<TOutputs>,
 > {
   readonly name?: string;
   readonly description: string;
@@ -73,7 +83,7 @@ export interface AuthoredWorkflowSpec<
   readonly worktreeFromInputs?: WorkflowWorktreeInputBinding;
   readonly run: (
     ctx: WorkflowRunContext<WorkflowInputsFromSchemas<TInputs>, WorkflowOutputsFromSchemas<TOutputs>>,
-  ) => Promise<WorkflowOutputsFromSchemas<TOutputs>> | WorkflowOutputsFromSchemas<TOutputs>;
+  ) => Promise<WorkflowRunOutputResult<TOutputs, TActualOutputs>> | WorkflowRunOutputResult<TOutputs, TActualOutputs>;
 }
 
 export type AuthoredWorkflowDefinition<
@@ -153,10 +163,11 @@ function freezeInputBindings(
 }
 
 export function workflow<
-  TInputs extends WorkflowInputSchemaMap,
-  TOutputs extends WorkflowOutputSchemaMap,
+  const TInputs extends WorkflowInputSchemaMap,
+  const TOutputs extends WorkflowOutputSchemaMap,
+  TActualOutputs extends WorkflowDeclaredOutputsFromSchemas<TOutputs> = WorkflowDeclaredOutputsFromSchemas<TOutputs>,
 >(
-  spec: AuthoredWorkflowSpec<TInputs, TOutputs>,
+  spec: AuthoredWorkflowSpec<TInputs, TOutputs, TActualOutputs>,
 ): AuthoredWorkflowDefinition<TInputs, TOutputs> {
   if (typeof spec.description !== "string") {
     throw new TypeError("workflow: description must be a string");
