+++ packages/workflows/src/engine/primitives/exit.ts
@@ -1,0 +1,2 @@
+export { createWorkflowExitManager } from "../../runs/foreground/executor-exit-manager.js";
+export type { WorkflowExitManager } from "../../runs/foreground/executor-exit-manager.js";
