+++ packages/workflows/src/engine/replay.ts
@@ -1,0 +1,8 @@
+export {
+  createContinuationReplayIndex,
+  sameStringSet,
+} from "../runs/foreground/executor-continuation.js";
+export type {
+  ContinuationReplayDecision,
+  ContinuationReplayIndex,
+} from "../runs/foreground/executor-continuation.js";
