+++ packages/workflows/src/sdk-surface.ts
@@ -28,8 +28,8 @@ export type { WorkflowRegistry } from "./workflows/registry.js";
 export { run, runTask, runParallel, runChain, resolveInputs } from "./runs/foreground/executor.js";
 export type { RunOpts, RunResult, ResolvedInputs } from "./runs/foreground/executor.js";
 export type { AgentSessionAdapter, StageAdapters } from "./runs/foreground/stage-runner.js";
-export { GraphFrontierTracker } from "./runs/shared/graph-inference.js";
-export type { StageNode } from "./runs/shared/graph-inference.js";
+export { GraphFrontierTracker } from "./engine/graph-inference.js";
+export type { StageNode } from "./engine/graph-inference.js";
 export { setupGitWorktree } from "./runs/shared/worktree.js";
 export type { GitWorktreeSetupOptions, GitWorktreeSetupResult } from "./runs/shared/worktree.js";
 export { createStore, store } from "./shared/store.js";
